Got conversions routines working in reverse for converting from ms formats to gui friendly versions.  Seems to work well.  update_ve_const() function is about 1/4 of its original size thanks to this new method... Still awaiting DualTable clarifications for DT maintainer(s) so I can implement support for the data on table two.

+++ src/conversions.h
@@ -22,7 +22,7 @@
 struct Conversion_Chart
 {
 	gchar *conv_type[128];
-	gint conv_factor[128];
+	gfloat conv_factor[128];
 };
 
 #endif
